Imp: 上書きモードでの矩形選択入力で選択範囲を削除するオプションを追加
[patchunicode:#821]
旧仕様に戻すためのオプションを追加します。
original 削除しない
sakura ほぼ当初から 削除する
r2611-現在 削除しない

http://sakura-editor.sourceforge.net/cgi-bin/cyclamen/cyclamen.cgi?log=data&v=7769
[patchunicode:#312]
[r2611]

git-svn-id: https://svn.code.sf.net/p/sakura-editor/code/sakura/trunk2@3845 f7ce1907-e4c7-47ca-9f76-12c87ed2c91c

+++ sakura_core/cmd/CViewCommander_Edit_advanced.cpp
@@ -127,17 +127,18 @@ void CViewCommander::Command_INDENT( const wchar_t* const pData, const CLogicInt
 	/* 矩形範囲選択中か */
 	if( m_pCommanderView->GetSelectionInfo().IsBoxSelecting() ){
 // 2012.10.31 Moca 上書きモードのときの選択範囲削除をやめる
-#if 0
-		// From Here 2001.12.03 hor
-		/* 上書モードのときは選択範囲削除 */
-		if( ! m_pCommanderView->IsInsMode() /* Oct. 2, 2005 genta */){
-			sSelectOld = GetSelect();
-			m_pCommanderView->DeleteData( false );
-			GetSelect() = sSelectOld;
-			m_pCommanderView->GetSelectionInfo().SetBoxSelect(true);
+// 2014.06.02 Moca 旧仕様を選べるように
+		if( GetDllShareData().m_Common.m_sEdit.m_bOverWriteBoxDelete ){
+			// From Here 2001.12.03 hor
+			/* 上書モードのときは選択範囲削除 */
+			if( ! m_pCommanderView->IsInsMode() /* Oct. 2, 2005 genta */){
+				sSelectOld = GetSelect();
+				m_pCommanderView->DeleteData( false );
+				GetSelect() = sSelectOld;
+				m_pCommanderView->GetSelectionInfo().SetBoxSelect(true);
+			}
+			// To Here 2001.12.03 hor
 		}
-		// To Here 2001.12.03 hor
-#endif
 
 		/* 2点を対角とする矩形を求める */
 		CLayoutRange rcSel;
